Move null effect to effect.h.
